+++ js/server.js
@@ -22,19 +22,19 @@ export async function sendDataToServer(formData) {
   }
 }
 
-export function blockSubmitButton(button) {
-  if (button) {
-    button.disabled = true;
-    button.classList.add('loading');
-  }
-}
+// export function blockSubmitButton(button) {
+//   if (button) {
+//     button.disabled = true;
+//     button.classList.add('loading');
+//   }
+// }
 
-export function unblockSubmitButton(button) {
-  if (button) {
-    button.disabled = false;
-    button.classList.remove('loading');
-  }
-}
+// export function unblockSubmitButton(button) {
+//   if (button) {
+//     button.disabled = false;
+//     button.classList.remove('loading');
+//   }
+// }
 
 export function createFormData(form) {
   return new FormData(form);
